requirements.txt
faiss-cpu
Flask==1.1.1
Flask-Cors==3.0.8
pandas==0.25.0
torch==1.2.0
transformers==3.3.0
xlrd
